+++ platform/native/windows-msvc/tt_spinlock_native.c
